Initial commit
This project builds upon another similarly named project of mine that provided the same functionality through a TUI. This could become a branch in the repository of that project but it has now diverged so much from the original that I believe it deserves its own repository.

Let it rip!